swaps column_type with line_type and fixes unterminated comment error
source_coordinate{column, line} has become confusing, since it is
printed the other way around (`{line, column}`). This was swapped to
reflect this.

There was also an unterminated comment error, where the EOF character
was reported to be where the comment started, due to a loss of
information. The new solution preserves both the start of the comment
(to report where the comment starts), and the end of the comment (to
report where the EOF token is).